AI 스킬 로직 추가

+++ Client/Assets/Scripts/Datas/Parser/CharacterStatTableFolder/CharacterStatTable.Helper.cs
@@ -4,7 +4,7 @@ public partial class CharacterStatTable : ScriptParser
 {
 	public CharacterStat GetStat(ENUM_ENTITY_TYPE type)
 	{
-		if (characterStatDictionary.TryGetValue((int)type, out var stat))
+		if (characterStatDictionary.TryGetValue(type, out var stat))
 		{
 			return stat;
 		}
